pyproject.toml
[tool.poetry]
name = "python-test"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.11"
pyyaml = "6.*"
account = {git = "https://github.com/ggramal/dos14-gramovich-git-flow.git", rev = "master"}
flask = "*"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.0"
black = "^23.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
